\name{pdf_plot}
\alias{pdf_plot}
\title{
  Plot classification to pdf file
}
\description{
  Plot classification to pdf file, used for testing the code.
}
\usage{
  pdf_plot(df, tz, pdffile, statenames)
}
\arguments{
  \item{df}{
    Data.frame with GPS time series and all classifications
  }
  \item{tz}{
    Character to specify timezone (timezone database name)
  }
  \item{pdffile}{
    Character to specify path to the to be created pdf file
  }
  \item{statenames}{
    ...
  }
}
\value{
  No value is created only a pdf file
}
