\name{pipeline}
\alias{pipeline}
\title{
  Central pipeline to process GPS and optionally merge in GGIR output
}
\description{
  Central pipeline to process GPS and optionally merge in GGIR output
}
\usage{
  pipeline(params, GGIRpath, outputDir)
}
\arguments{
  \item{params}{
    Data.frame with all parameters needed in the pipeline
  }
  \item{GGIRpath}{
    Data.frame with all parameters needed in the pipeline
  }
   \item{outputDir}{
    Data.frame with all parameters needed in the pipeline
  }
}
\value{
  Data.frame with GPS time series and added classifications
}
